fix(sdk-go): address review findings across the SDK, client and instrumentations
Resolves 95 CodeRabbit review threads on the rebased branch. Six were
correctness bugs rather than polish:

- Bedrock's stream accumulator released its lock before reading the
  accumulator, racing the pump goroutine. Reproduces under -race before
  the fix.
- The REST client shared one math/rand source across concurrent requests.
- otelhttp marked a span Ok on 200 response headers, and OTel refuses to
  downgrade Ok, so a mid-stream provider error exported as a successful
  generation. Anthropic and OpenAI were both affected.
- Gemini's reasoning tokens were never priced: candidatesTokenCount
  excludes thoughtsTokenCount, so they were recorded as informational.
- DataCaptureNone leaked content twice over: langwatch.rag.contexts
  carries retrieved document text and was not stripped, and the filter
  covered attributes only, so free-text feedback in span events passed
  through untouched.
- go:generate pointed at an OpenAPI path the restructure had moved.

Eight findings were rejected with reasoning on the threads, most of them
resting on a house style the repository does not hold or on a semconv
package layout that does not exist.

The Go SDK's spec scenarios now bind for the first time: sdks/go joins
the feature-parity checker's Go roots, taking
specs/go-sdk/span-attribute-parity.feature from 6 of 19 enforced to
20 of 20.

BREAKING CHANGE: the provider middlewares now capture input and output
content by default (langwatch.DataCaptureAll), and the opt-in
WithCaptureInput() / WithCaptureOutput() options are removed. An
existing user who upgrades without passing WithDataCapture will begin
exporting full prompts and completions, which routinely contain end-user
PII. Opt out with WithDataCapture(langwatch.DataCaptureNone).

BREAKING CHANGE: the REST client no longer retries non-idempotent
requests. Retries now apply to RFC 9110 idempotent methods, plus POST
and PATCH carrying an Idempotency-Key. Evaluations.Create and
Events.Track are POSTs and are no longer retried on 429/5xx, because a
retry there duplicates an evaluation or a tracked event.

+++ sdks/go/instrumentation/ollama/extractor_generic.go
@@ -195,7 +195,10 @@ func metricsFromMap(body otelhttp.JSONObject) metricsPayload {
 }
 
 // mergeFromMap folds any present token / duration fields from an untyped body
-// into m, overwriting only when a non-zero value is found.
+// into m, overwriting only when a non-zero value is found. The durations are
+// nanoseconds on the Ollama wire — a one-second call is already 1e9 and a
+// multi-second call exceeds the int32 range — so they are read through GetInt64,
+// which does not truncate on 32-bit builds (GOARCH=386, 32-bit ARM).
 func (m *metricsPayload) mergeFromMap(body otelhttp.JSONObject) {
 	if v, ok := otelhttp.GetInt(body, "prompt_eval_count"); ok && v > 0 {
 		m.PromptEvalCount = v
@@ -203,16 +206,16 @@ func (m *metricsPayload) mergeFromMap(body otelhttp.JSONObject) {
 	if v, ok := otelhttp.GetInt(body, "eval_count"); ok && v > 0 {
 		m.EvalCount = v
 	}
-	if v, ok := otelhttp.GetInt(body, "total_duration"); ok && v > 0 {
-		m.TotalDuration = int64(v)
+	if v, ok := otelhttp.GetInt64(body, "total_duration"); ok && v > 0 {
+		m.TotalDuration = v
 	}
-	if v, ok := otelhttp.GetInt(body, "load_duration"); ok && v > 0 {
-		m.LoadDuration = int64(v)
+	if v, ok := otelhttp.GetInt64(body, "load_duration"); ok && v > 0 {
+		m.LoadDuration = v
 	}
-	if v, ok := otelhttp.GetInt(body, "prompt_eval_duration"); ok && v > 0 {
-		m.PromptEvalDuration = int64(v)
+	if v, ok := otelhttp.GetInt64(body, "prompt_eval_duration"); ok && v > 0 {
+		m.PromptEvalDuration = v
 	}
-	if v, ok := otelhttp.GetInt(body, "eval_duration"); ok && v > 0 {
-		m.EvalDuration = int64(v)
+	if v, ok := otelhttp.GetInt64(body, "eval_duration"); ok && v > 0 {
+		m.EvalDuration = v
 	}
 }
